Refactor stringifiers
